+++ R/polca-tidiers.R
@@ -115,26 +115,14 @@ tidy.poLCA <- function(x, ...) {
 #' @family poLCA tidiers
 augment.poLCA <- function(x, data = NULL, ...) {
   indices <- cbind(seq_len(nrow(x$posterior)), x$predclass)
-  ret <- tibble(
+  
+  new_cols <- tibble(
     .class = x$predclass,
     .probability = x$posterior[indices]
   )
 
-  if (is.null(data)) {
-    data <- x$y
-    if (!is.null(x$x)) {
-      data <- cbind(data, x$x)
-    }
-  } else {
-    if (nrow(data) != nrow(ret)) {
-      # rows may have been removed for NAs.
-      # For those rows, the new columns get NAs
-      ret$.rownames <- rownames(x$y)
-      ret <- ret[rownames(data), ]
-    }
-  }
-
-  as_tibble(cbind(data, ret))
+  data <- if (is.null(data)) bind_cols(x$y, x$x) else as_broom_tibble(data)
+  as_tibble(bind_cols(data, new_cols))
 }
 
 #' @templateVar class poLCA