+++ tests/testthat/test-polca.R
@@ -30,13 +30,11 @@ test_that("glance.poLCA", {
 test_that("augment.poLCA", {
   
   au <- augment(fit)
-  check_tibble(au, method = "augment", strict = FALSE)
   
   check_augment_function(
     aug = augment.poLCA,
     model = fit,
-    data = values,
-    strict = FALSE
+    data = values
   )
 })